refactor(codex): extract thread lifecycle preflight
