+++ extensions/codex/src/app-server/thread-lifecycle-run.ts
@@ -1,32 +1,17 @@
-import {
-  embeddedAgentLog,
-  formatErrorMessage,
-  isHostScopedAgentToolActive,
-} from "openclaw/plugin-sdk/agent-harness-runtime";
-import { buildCodexUserMcpServersThreadConfigPatchForRuntime } from "openclaw/plugin-sdk/codex-mcp-projection";
+import { embeddedAgentLog } from "openclaw/plugin-sdk/agent-harness-runtime";
 import { isIncognitoSessionKey } from "../incognito-session.js";
 import { closeCodexStartupClientBestEffort } from "./attempt-client-cleanup.js";
-import {
-  getCodexAppServerClientInstanceId,
-  resolveCodexAppServerClientInstanceId,
-} from "./client.js";
-import { isSystemAgentOnlyCodexDynamicToolAllowlist } from "./dynamic-tool-profile.js";
-import {
-  applyCodexNativeSkillIsolation,
-  resolveCodexNativeSkillIsolation,
-} from "./native-skill-isolation.js";
+import { resolveCodexAppServerClientInstanceId } from "./client.js";
+import { applyCodexNativeSkillIsolation } from "./native-skill-isolation.js";
 import { buildCodexAppServerConnectionFingerprint } from "./plugin-app-cache-key.js";
 import {
   isCodexPluginThreadBindingStale,
   mergeCodexThreadConfigs,
   type CodexPluginThreadConfig,
 } from "./plugin-thread-config.js";
-import { isCodexAppServerProfilerEnabled } from "./profiler-flag.js";
-import { flattenCodexDynamicToolFunctions } from "./protocol.js";
 import {
   assertCodexBindingMayBeReplaced,
   createCodexSessionGenerationSupersededError,
-  hashCodexAppServerBindingFingerprint,
   normalizeCodexAppServerBindingModelProvider,
   reclaimCurrentCodexSessionGeneration,
   sessionBindingIdentity,
@@ -40,23 +25,15 @@ import {
   shouldRotateCodexAppServerBindingForRuntime,
   shouldRotateCodexGpt56MultiAgentBinding,
 } from "./thread-binding-policy.js";
-import {
-  buildContextEngineBinding,
-  isContextEngineBindingCompatible,
-} from "./thread-context-engine.js";
+import { isContextEngineBindingCompatible } from "./thread-context-engine.js";
 import {
   areDynamicToolFingerprintsCompatible,
   areUserMcpServersFingerprintsCompatible,
-  codexLegacyDynamicToolsFingerprint as legacyFingerprintDynamicTools,
-  fingerprintEnvironmentSelection,
-  fingerprintJsonObject,
-  fingerprintUserMcpServersConfigPatch,
-  legacyFingerprintUserMcpServersConfigPatch,
   shouldStartTransientNoToolThread,
 } from "./thread-fingerprints.js";
 import { CodexThreadBindingConflictError } from "./thread-lifecycle-errors.js";
 import { resumeExistingCodexThread, startFreshCodexThread } from "./thread-lifecycle-io.js";
-import { createCodexThreadLifecycleTimingTracker } from "./thread-lifecycle-timing.js";
+import { prepareCodexThreadLifecyclePreflight } from "./thread-lifecycle-preflight.js";
 import type {
   CodexAppServerThreadLifecycleBinding,
   CodexStartOrResumeThreadParams,
@@ -68,14 +45,7 @@ import {
   tryReuseCodexLiveThread,
 } from "./thread-lifecycle-warm.js";
 import { resolveCodexAppServerThreadModelSelection } from "./thread-model-selection.js";
-import {
-  assertCodexRingZeroHasNoManagedHooks,
-  buildCodexRingZeroThreadConfigPatch,
-  CODEX_RING_ZERO_BASE_INSTRUCTIONS,
-  readCodexInheritedMcpServerNames,
-} from "./thread-requests.js";
 import { materializePendingSupervisionBranch } from "./thread-supervision.js";
-import { resolveCodexWebSearchPlan } from "./web-search.js";
 
 export async function startOrResumeThread(
   params: CodexStartOrResumeThreadParams,
@@ -89,103 +59,26 @@ export async function startOrResumeThread(
     config: params.params.config,
   });
   return await params.bindingStore.withLease(bindingIdentity, async () => {
-    // Thread lifecycle spans are useful when profiling startup churn, but normal
-    // turns should not pay Date.now/span-array overhead while resuming threads.
-    const lifecycleTiming = createCodexThreadLifecycleTimingTracker({
-      ...params.timing,
-      enabled: params.timing?.enabled ?? isCodexAppServerProfilerEnabled(params.params.config),
-    });
-    const legacyDynamicToolsFingerprint = lifecycleTiming.measureSync(
-      "legacy-dynamic-tools-fingerprint",
-      () => legacyFingerprintDynamicTools(params.dynamicTools),
-    );
-    const dynamicToolsFingerprint = lifecycleTiming.measureSync("dynamic-tools-fingerprint", () =>
-      hashCodexAppServerBindingFingerprint(legacyDynamicToolsFingerprint),
-    );
-    const dynamicToolsContainDeferred = flattenCodexDynamicToolFunctions(params.dynamicTools).some(
-      (tool) => tool.deferLoading === true,
-    );
-    const webSearchPlan = lifecycleTiming.measureSync("web-search-plan", () =>
-      resolveCodexWebSearchPlan({
-        config: params.params.config,
-        disableTools: params.params.disableTools,
-        nativeToolSurfaceEnabled: params.nativeCodeModeEnabled,
-        nativeProviderWebSearchSupport: params.nativeProviderWebSearchSupport,
-        webSearchAllowed: params.webSearchAllowed,
-      }),
-    );
-    const webSearchThreadConfigFingerprint = fingerprintJsonObject(webSearchPlan.threadConfig);
-    const networkProxyConfigFingerprint = params.appServer.networkProxy?.configFingerprint;
-    const contextEngineBinding = lifecycleTiming.measureSync("context-engine-binding", () =>
-      buildContextEngineBinding(params.params, params.contextEngineProjection),
-    );
-    const userMcpServersConfigPatch =
-      params.userMcpServersEnabled === false
-        ? undefined
-        : await buildCodexUserMcpServersThreadConfigPatchForRuntime(params.params.config, {
-            agentId: params.agentId ?? params.params.agentId,
-            agentDir: params.params.agentDir,
-            allowLiteralOAuthProjection: params.appServer.connectionClass !== "remote",
-            onServerUnavailable: (serverName, error) =>
-              embeddedAgentLog.warn("skipping unavailable MCP OAuth server", {
-                serverName,
-                error: formatErrorMessage(error),
-              }),
-          });
-    const nativeSkillIsolation = await lifecycleTiming.measure("native-skill-isolation", () =>
-      resolveCodexNativeSkillIsolation({
-        client: params.client,
-        codexHome: params.appServer.start.env?.CODEX_HOME,
-        cwd: params.cwd,
-        home: params.appServer.start.env?.HOME,
-        signal: params.signal,
-        userProfile: params.appServer.start.env?.USERPROFILE,
-      }),
-    );
-    const nativeSkillIsolationFingerprint = nativeSkillIsolation
-      ? fingerprintJsonObject({
-          version: 1,
-          disabledUserSkillPaths: nativeSkillIsolation.disabledUserSkillPaths,
-        })
-      : undefined;
-    const legacyUserMcpServersFingerprint =
-      legacyFingerprintUserMcpServersConfigPatch(userMcpServersConfigPatch);
-    const userMcpServersFingerprint =
-      fingerprintUserMcpServersConfigPatch(userMcpServersConfigPatch);
-    const environmentSelectionFingerprint = fingerprintEnvironmentSelection(
-      params.environmentSelection,
-    );
-    const hostSystemAgentActive =
-      params.hostSystemAgentActive ?? isHostScopedAgentToolActive("openclaw");
-    const ringZeroActive =
-      hostSystemAgentActive && isSystemAgentOnlyCodexDynamicToolAllowlist(params.params.toolsAllow);
-    if (ringZeroActive && params.nativeCodeModeEnabled !== false) {
-      throw new Error("Codex ring-zero requires native code mode to be disabled");
-    }
-    const ringZeroInheritedMcpServerNames = ringZeroActive
-      ? await lifecycleTiming.measure("ring-zero-mcp-config-read", () =>
-          readCodexInheritedMcpServerNames(params.client, params.cwd, params.signal),
-        )
-      : [];
-    if (ringZeroActive) {
-      await lifecycleTiming.measure("ring-zero-config-requirements-read", () =>
-        assertCodexRingZeroHasNoManagedHooks(params.client, params.signal),
-      );
-    }
-    const ringZeroConfigFingerprint = ringZeroActive
-      ? fingerprintJsonObject({
-          version: 1,
-          baseInstructions: CODEX_RING_ZERO_BASE_INSTRUCTIONS,
-          config: buildCodexRingZeroThreadConfigPatch(
-            params.params,
-            true,
-            ringZeroInheritedMcpServerNames,
-          )!,
-        })
-      : undefined;
-    const ringZeroClientInstanceId = ringZeroActive
-      ? getCodexAppServerClientInstanceId(params.client)
-      : undefined;
+    const {
+      contextEngineBinding,
+      dynamicToolsContainDeferred,
+      dynamicToolsFingerprint,
+      environmentSelectionFingerprint,
+      hostSystemAgentActive,
+      legacyDynamicToolsFingerprint,
+      legacyUserMcpServersFingerprint,
+      lifecycleTiming,
+      nativeSkillIsolation,
+      nativeSkillIsolationFingerprint,
+      networkProxyConfigFingerprint,
+      ringZeroActive,
+      ringZeroClientInstanceId,
+      ringZeroConfigFingerprint,
+      ringZeroInheritedMcpServerNames,
+      userMcpServersConfigPatch,
+      userMcpServersFingerprint,
+      webSearchThreadConfigFingerprint,
+    } = await prepareCodexThreadLifecyclePreflight(params);
     let binding = await lifecycleTiming.measure("read-binding", () =>
       params.bindingStore.read(bindingIdentity),
     );
